object_move_pipeline/requirements.txt
# Object Move Pipeline Requirements
# Python 3.8+

# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0

# SAM (Segment Anything Model)
transformers>=4.30.0
accelerate>=0.20.0

# ObjectClear Pipeline
diffusers>=0.21.0
omegaconf>=2.3.0

# AnyDoor Dependencies
einops>=0.6.0
pytorch-lightning>=2.0.0
albumentations>=1.3.0

# Image Processing
scikit-image>=0.19.0
imageio>=2.25.0

# Evaluation Metrics (Optional but Recommended)
lpips>=0.1.4
pytorch-fid>=0.3.0
image-quality>=1.2.7

# Utilities
tqdm>=4.65.0
matplotlib>=3.5.0

